normal-bot/requirements.txt
discord.py
Flask
python-dotenv
aiohttp